+++ CHANGELOG.md
@@ -18,7 +18,8 @@
   * added clone and save as template
   * added `NewTemplateGui`
   * added template selection to `InsertGui`
-  * (X) added template editing and deleting
+  * added template editing and deleting
+  * (X) import and export templates
 
 ## v0.8.4
 * minor bug fixes